Adding Agenda app partially done

+++ Agenda/app/src/main/java/com/example/agenda/User.kt
@@ -1,3 +1,3 @@
 package com.example.agenda
 
-data class User constructor(var name: String, var address: String, var phone: String, var contactType: UByte){}
+data class User constructor(var name: String, var address: String, var phone: String, var contactType: Int){}
